Market price methods to operate on condensed data
Instead of operating on `[u64]`, these new methods will instead operate
on `[(u64, u64)]` where the tuple represents a `(price, quantity)` pair,
which is the data that is present on any one given auction.

While these methods (specifically normalized_market_price) is
approximately 25% slower than the `[u64]` variant, a significant chunk
of performance is gained by not needing to perform a large number of
vector allocations in order to generate the full price array.